+++ email_auth/conftest.py
@@ -6,6 +6,15 @@ from email_auth import models
 
 
 @pytest.fixture
+def mock_email_address_qs():
+    mock_qs = mock.Mock(spec=models.EmailAddress.objects)
+    mock_qs.all.return_value = mock_qs
+
+    with mock.patch("email_auth.models.EmailAddress.objects", new=mock_qs):
+        yield mock_qs
+
+
+@pytest.fixture
 def mock_email_verification_qs():
     mock_qs = mock.Mock(spec=models.EmailVerification.objects)
     mock_qs.all.return_value = mock_qs